// src/types/employee.ts
export interface Employee {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  position: string;
  department: string;
  phoneNumber: string;
  hireDate: string;
}

export interface EmployeeFormData {
  firstName: string;
  lastName: string;
  email: string;
  position: string;
  department: string;
  phoneNumber: string;
  hireDate: string;
} 